% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_happifndata.R
\name{check_happifndata}
\alias{check_happifndata}
\title{check_happifndata}
\usage{
check_happifndata(version = "0.0.0.9000")
}
\arguments{
\item{version}{\code{character}; version of \code{happifndata} to check}
}
\description{
check_happifndata
}
\examples{
\dontrun{
check_happifndata()
}

}
