% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_names}
\alias{get_dataset_names}
\title{get_dataset_names}
\usage{
get_dataset_names()
}
\value{
\code{character} containing all dataset name available from IFN
}
\description{
get_dataset_names
}
\examples{
\dontrun{
get_dataset_names()
get_dataset_names()[1]
}
}
