% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifn.R
\name{get_ifn}
\alias{get_ifn}
\title{get_ifn}
\usage{
get_ifn(name,
               dir = NULL,
               quiet = FALSE)
}
\arguments{
\item{name}{\code{character}; Name of needed dataset.}

\item{dir}{\code{character} or \code{NULL}; connection open for writing.
(ex : "~/test"). If \code{NULL}, dataset are seamlessly imported into memory.
Be careful; it may take a while.}

\item{quiet}{\code{boolean}; If TRUE no message are displayed.}
}
\value{
\code{NULL}; datasets are loaded intot emory
}
\description{
Load IFN dataset(s) into memory. If provided, dataset are loaded from a
directory, else they are seamlessly downloaded before loading.
}
\examples{
\dontrun{
arbre <- get_ifn(name = "arbre", dir = NULL)

all_dataset <- lapply(get_dataset_names(), get_ifn) |>
               setNames(get_dataset_names())
}
}
