% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifn_metadata.R
\name{get_ifn_metadata}
\alias{get_ifn_metadata}
\title{get_ifn_metadata}
\usage{
get_ifn_metadata(dir = NULL,
                    unlist = FALSE,
                    quiet = FALSE)
}
\arguments{
\item{dir}{\code{character} or \code{NULL}; connection open for writing.
(ex : "~/test"). If \code{NULL}, dataset are seamlessly imported into memory.
Be careful; it may take a while.}

\item{unlist}{\code{boolean}; if TRUE, all datasets are load into globalenv. if
FALSE, a list of datasets is returned.}

\item{quiet}{\code{boolean}; If TRUE no message are displayed.}
}
\value{
\code{list}; three dataset
}
\description{
Load a list with three metadata dataset
}
\examples{
\dontrun{
TO-DO
}
}
