% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(dir = NULL)
}
\arguments{
\item{dir}{\code{character} or \code{NULL}; connection open for writing.
(ex : "~/test"). If \code{NULL}, dataset are seamlessly imported into memory.
Be careful; it may take a while.}
}
\value{
\code{list}; three dataset
}
\description{
Load a list with three metadata dataset
}
\examples{
\dontrun{
TO-DO
}
}
