% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{get_url}
\usage{
get_url(all = F)
}
\arguments{
\item{all}{\code{logical}; If \code{TRUE} url for downloading data since 2005 is returned.
Else, only data from 2021 is returned.}
}
\value{
.zip url for downloading data
}
\description{
get_url
}
\examples{
\dontrun{
get_url()
}
}
