% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifn.R
\name{get_ifn}
\alias{get_ifn}
\title{get_ifn}
\usage{
get_ifn(names = "arbre", dir = NULL)
}
\arguments{
\item{names}{\code{character}; Name of needed dataset.}

\item{dir}{\code{character} or \code{NULL}; connection open for writing.
(ex : "~/test"). If \code{NULL}, dataset are seamlessly imported into memory.
Be careful; it may take a while.}
}
\value{
\code{NULL}; datasets are loaded intot emory
}
\description{
Load IFN dataset(s) into memory. If provided, dataset are loaded from a
directory, else they are seamlessly downloaded before loading.
}
\examples{
\dontrun{
all_dataset <- get_ifn(names = get_dataset_names())
}
}
