% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifn.R, R/utils.R
\name{get_ifn_all}
\alias{get_ifn_all}
\title{get_ifn_all}
\usage{
get_ifn_all(unlist = TRUE)

get_ifn_all(unlist = TRUE)
}
\arguments{
\item{unlist}{\code{boolean}; if TRUE, all datasets are load into globalenv. if
FALSE, a list of datasets is returned.}
}
\value{
\code{NULL} or \code{list}

\code{NULL} or \code{list}
}
\description{
Wrapper function to load all ifn datasets

Wrapper function to load all ifn datasets
}
\examples{
\dontrun{
# all ifn data as lis
ifn_data <- get_ifn_all()

# directly load data into globalenv
get_ifn_all(unlist = TRUE)
}

}
